refactor(test): 테스트 유틸함수 디렉토리 변경
src root에서 src/utils/test로 이동
