+++ __tests__/ApplicationTest.js
@@ -1,6 +1,6 @@
 import { MissionUtils } from "@woowacourse/mission-utils";
 import App from "../src/App.js";
-import getLogSpy from "../src/getLogSpy.js";
+import getLogSpy from "../src/utils/test/getLogSpy.js";
 
 const mockQuestions = (inputs) => {
   MissionUtils.Console.readLineAsync = jest.fn();
